+++ lib/jbr/mock/visit.rb
@@ -15,11 +15,9 @@ module Jbr
 
     def job_id = @node[:job_id]
 
-    def property_id = @node[:property_id]
-
     def client = @node.fetch :client, {}
 
-    def address = @node.fetch :address, {}
+    def property = @node.fetch :property, {}
 
     def all_day? = @node[:all_day]